mc1/requirements.txt
asgiref==3.4.1
certifi==2021.5.30
Django==3.2.2
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.12.4
channels==3.0.4
channels-redis==3.3.0
